Move MethodDefinition
