+++ src/ClassDefinition/ClassDefinition.php
@@ -6,7 +6,7 @@ namespace webignition\BasilCompilationSource\ClassDefinition;
 
 use webignition\BasilCompilationSource\Metadata\Metadata;
 use webignition\BasilCompilationSource\Metadata\MetadataInterface;
-use webignition\BasilCompilationSource\MethodDefinitionInterface;
+use webignition\BasilCompilationSource\MethodDefinition\MethodDefinitionInterface;
 use webignition\BasilCompilationSource\SourceInterface;
 
 class ClassDefinition implements ClassDefinitionInterface
@@ -14,7 +14,7 @@ class ClassDefinition implements ClassDefinitionInterface
     private $name;
 
     /**
-     * @var MethodDefinitionInterface[]
+     * @var \webignition\BasilCompilationSource\MethodDefinition\MethodDefinitionInterface[]
      */
     private $methods = [];
